refactor(python): rename `read_sql` to `read_database` (#7315)
