+++ py-polars/polars/__init__.py
@@ -147,6 +147,7 @@ from polars.io import (
     read_avro,
     read_csv,
     read_csv_batched,
+    read_database,
     read_delta,
     read_excel,
     read_ipc,
@@ -219,6 +220,7 @@ __all__ = [
     "read_avro",
     "read_csv",
     "read_csv_batched",
+    "read_database",
     "read_delta",
     "read_excel",
     "read_ipc",
