+++ py-polars/polars/io/__init__.py
@@ -2,6 +2,7 @@
 
 from polars.io.avro import read_avro
 from polars.io.csv import read_csv, read_csv_batched, scan_csv
+from polars.io.database import read_database, read_sql
 from polars.io.delta import read_delta, scan_delta
 from polars.io.excel import read_excel
 from polars.io.ipc import read_ipc, scan_ipc
@@ -9,12 +10,12 @@ from polars.io.json import read_json
 from polars.io.ndjson import read_ndjson, scan_ndjson
 from polars.io.parquet import read_parquet, scan_parquet
 from polars.io.pyarrow_dataset import scan_ds
-from polars.io.sql import read_sql
 
 __all__ = [
     "read_avro",
     "read_csv",
     "read_csv_batched",
+    "read_database",
     "read_delta",
     "read_excel",
     "read_ipc",
